requirements.txt
discord.py>=2.3.0
feedparser>=6.0.10
googletrans==4.0.0rc1
python-dotenv>=1.0.0
schedule>=1.2.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
